+++ app/post/burn.go
@@ -7,6 +7,8 @@ import (
 	errorsmod "cosmossdk.io/errors"
 	sdkmath "cosmossdk.io/math"
 
+	"github.com/armon/go-metrics"
+	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
@@ -54,7 +56,7 @@ func (bd BurnDecorator) PostHandle(ctx sdk.Context, tx sdk.Tx, simulate, success
 	}
 
 	// burn min(balance, fee)
-	var burnedCoins sdk.Coins
+	var burntCoins sdk.Coins
 	for _, fee := range fees {
 		balance := bd.bankKeeper.GetBalance(ctx, authtypes.NewModuleAddress(bd.feeCollectorName), fee.Denom)
 		if !balance.IsPositive() {
@@ -63,14 +65,30 @@ func (bd BurnDecorator) PostHandle(ctx sdk.Context, tx sdk.Tx, simulate, success
 
 		amount := sdkmath.MinInt(fee.Amount, balance.Amount)
 
-		burnedCoins = append(burnedCoins, sdk.Coin{Denom: fee.Denom, Amount: amount})
+		burntCoins = append(burntCoins, sdk.Coin{Denom: fee.Denom, Amount: amount})
 	}
 
 	// NOTE: since all Cosmos tx fees are pooled by the fee collector module account,
 	// we burn them directly from it
-	if err := bd.bankKeeper.BurnCoins(ctx, bd.feeCollectorName, burnedCoins); err != nil {
+	if err := bd.bankKeeper.BurnCoins(ctx, bd.feeCollectorName, burntCoins); err != nil {
 		return ctx, err
 	}
+
+	defer func() {
+		for _, c := range burntCoins {
+			// if fee amount is higher than uint64, skip the counter
+			if !c.Amount.IsUint64() {
+				continue
+			}
+			telemetry.IncrCounterWithLabels(
+				[]string{"burnt", "tx", "fee", "amount"},
+				float32(c.Amount.Uint64()),
+				[]metrics.Label{
+					telemetry.NewLabel("denom", c.Denom),
+				},
+			)
+		}
+	}()
 
 	return next(ctx, tx, simulate, success)
 }