+++ CHANGELOG.md
@@ -154,6 +154,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (p256-precompile) [#2228](https://github.com/evmos/evmos/pull/2228) Adjust p256 precompile address from `0x0b` to `0x100`.
 - (staking-precompile) [#2234](https://github.com/evmos/evmos/pull/2234) Fix wrong error messages in `NewMsgCreateValidator`.
 - (debug) [#2245](https://github.com/evmos/evmos/pull/2245) Update convert an address between hex and bech32.
+- (metrics) [#2246](https://github.com/evmos/evmos/pull/2246) Add burnt cosmos transactions fees metric.
 
 ### Bug Fixes